refactored prints and plots
