+++ run_tests.py
@@ -9,5 +9,5 @@ def run():
     # def run_nms_test():
     #     print('afsf')
     #     test_nms.test()
-    test_tensor_unfolding.test()
+    # test_tensor_unfolding.test()
 run()
